+++ CHANGELOG.md
@@ -32,6 +32,7 @@ system. Old cached data will be converted to the new format automatically.
 * Always quote If-None-Match headers (#584)
 * Widget only pulses when syncing changes (#437)
 * Fire `connected` event on startup if remote already connected
+* Fix error on remoteStorage.local.reset (#630)
 
 ## 0.9.0 (December 2013)